chore(tests): move functions unit tests into src folder
