+++ .github/workflows/app-deploy.yml
@@ -171,7 +171,7 @@ jobs:
           echo Finalizing the release...
           $(yarn bin)/sentry-cli releases finalize $releaseName
           echo Marking new release as deployed...
-          $(yarn bin)/sentry-cli releases deploys $releaseName new -e $environment
+          $(yarn bin)/sentry-cli releases deploys $releaseName new -e $envName
           echo Release created successfully
 
       - name: Create Release
