+++ functions/src/cleanupProject/cleanupProject.spec.ts
@@ -2,7 +2,7 @@ import * as admin from 'firebase-admin'
 import { expect } from 'chai';
 import sinon from 'sinon';
 import functionsTestLib from 'firebase-functions-test'
-import { to } from '../../src/utils/async'
+import { to } from '../utils/async'
 
 const functionsTest = functionsTestLib()
